==> 1_Save 1-minute audio as 1-second segments.ipynb <==
import librosa
from glob import glob
from tqdm import tqdm
import numpy as np
import datetime

dic_list = glob(r'sound file/*')

dic_list

for dic_path in dic_list:

    file_list = [x for x in glob(dic_path+"/*.wav")]

    for file in tqdm(file_list):

        sig, sampling_rate = librosa.load(file, sr=44100)

        start_time = datetime.datetime.strptime(file[-23:-4], '%Y_%m_%d_%H_%M_%S')

        for sec in range(0,59):

            temp_sig = sig[(sec)*sampling_rate:(sec+1)*sampling_rate]

            time = start_time+datetime.timedelta(seconds=sec)
            time_str = time.strftime("%Y_%m_%d_%H_%M_%S")

            np.save(r'second/{}.npy'.format(time_str), temp_sig)



        #sec == 59인 경우 padding
        temp_sig = sig[59*sampling_rate:]
        padded_sig = np.concatenate((temp_sig, temp_sig[-4080:]))

        time = start_time+datetime.timedelta(seconds=59)
        time_str = time.strftime("%Y_%m_%d_%H_%M_%S")

        np.save(r'second/{}.npy'.format(time_str), padded_sig)
